refactor: user -> member 교체
