+++ src/main/java/com/example/bookrecommandations/common/exception/DuplicatedMembernameException.java
@@ -1,9 +1,9 @@
 package com.example.bookrecommandations.common.exception;
 
-public class DuplicatedUsernameException extends  RuntimeException{
+public class DuplicatedMembernameException extends  RuntimeException{
     private final ErrorCode errorCode;
 
-    public DuplicatedUsernameException(ErrorCode errorCode) {
+    public DuplicatedMembernameException(ErrorCode errorCode) {
         super(errorCode.getMessage());
         this.errorCode = errorCode;
     }
