+++ src/main/java/com/example/bookrecommandations/common/exception/ErrorCode.java
@@ -5,7 +5,7 @@ import org.springframework.http.HttpStatus;
 
 @Getter
 public enum ErrorCode {
-    DUPLICATED_USERNAME(HttpStatus.BAD_REQUEST, "이미 사용중인 아이디입니다."),
+    DUPLICATED_MEMBERNAME(HttpStatus.BAD_REQUEST, "이미 사용중인 아이디입니다."),
     DUPLICATED_NICKNAME(HttpStatus.BAD_REQUEST, "이미 사용중인 닉네임입니다.");
 
     private final HttpStatus status;
