+++ src/main/java/com/example/bookrecommandations/common/exception/GlobalExceptionHandler.java
@@ -1,6 +1,6 @@
 package com.example.bookrecommandations.common.exception;
 
-import com.example.bookrecommandations.user.dto.ErrorResponseDTO;
+import com.example.bookrecommandations.member.dto.ErrorResponseDTO;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
@@ -11,8 +11,8 @@ import org.springframework.web.bind.annotation.RestControllerAdvice;
 public class GlobalExceptionHandler {
 
     @ResponseStatus(HttpStatus.BAD_REQUEST)
-    @ExceptionHandler(DuplicatedUsernameException.class)
-    public ResponseEntity<ErrorResponseDTO> handleDeuplicatedUsernameException(DuplicatedUsernameException e) {
+    @ExceptionHandler(DuplicatedMembernameException.class)
+    public ResponseEntity<ErrorResponseDTO> handleDeuplicatedMembernameException(DuplicatedMembernameException e) {
         ErrorResponseDTO errorResponseDTO = new ErrorResponseDTO(HttpStatus.BAD_REQUEST.value(), e.getMessage());
 
         return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(errorResponseDTO);
