+++ src/main/java/com/example/bookrecommandations/member/controller/MemberController.java
@@ -1,8 +1,8 @@
-package com.example.bookrecommandations.user.controller;
+package com.example.bookrecommandations.member.controller;
 
-import com.example.bookrecommandations.user.dto.CreateMemberRequest;
-import com.example.bookrecommandations.user.dto.SaveResponseDTO;
-import com.example.bookrecommandations.user.service.MemberService;
+import com.example.bookrecommandations.member.dto.CreateMemberRequest;
+import com.example.bookrecommandations.member.dto.SaveResponseDTO;
+import com.example.bookrecommandations.member.service.MemberService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -17,11 +17,11 @@ import org.springframework.web.bind.annotation.RestController;
 @RequestMapping("/book")
 public class MemberController {
 
-    private final MemberService userService;
+    private final MemberService memberService;
 
     @PostMapping("/signup")
     public ResponseEntity<SaveResponseDTO> signup(@Validated @RequestBody CreateMemberRequest createMemberRequest) {
-        Long memberId = userService.saveMember(createMemberRequest);
+        Long memberId = memberService.saveMember(createMemberRequest);
         return ResponseEntity.ok(new SaveResponseDTO(
                 memberId, HttpStatus.OK.value(), "회원가입이 정상적으로 진행되었습니다."
         ));
