+++ src/main/java/com/example/bookrecommandations/member/domain/PreferredGenre.java
@@ -1,4 +1,4 @@
-package com.example.bookrecommandations.user.domain;
+package com.example.bookrecommandations.member.domain;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
