+++ src/main/java/com/example/bookrecommandations/member/domain/Survey.java
@@ -1,7 +1,7 @@
-package com.example.bookrecommandations.user.domain;
+package com.example.bookrecommandations.member.domain;
 
-import com.example.bookrecommandations.user.vo.Age;
-import com.example.bookrecommandations.user.vo.Gender;
+import com.example.bookrecommandations.member.vo.Age;
+import com.example.bookrecommandations.member.vo.Gender;
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
