+++ src/main/java/com/example/bookrecommandations/member/repository/PreferredBookRepository.java
@@ -1,6 +1,6 @@
-package com.example.bookrecommandations.user.repository;
+package com.example.bookrecommandations.member.repository;
 
-import com.example.bookrecommandations.user.domain.PreferredBook;
+import com.example.bookrecommandations.member.domain.PreferredBook;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface PreferredBookRepository extends JpaRepository<PreferredBook, Long> {
