+++ src/main/java/com/example/bookrecommandations/member/repository/PreferredGenreRepository.java
@@ -1,6 +1,6 @@
-package com.example.bookrecommandations.user.repository;
+package com.example.bookrecommandations.member.repository;
 
-import com.example.bookrecommandations.user.domain.PreferredGenre;
+import com.example.bookrecommandations.member.domain.PreferredGenre;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface PreferredGenreRepository extends JpaRepository<PreferredGenre, Long> {
