+++ src/main/java/com/example/bookrecommandations/member/repository/SurveyRepository.java
@@ -1,6 +1,6 @@
-package com.example.bookrecommandations.user.repository;
+package com.example.bookrecommandations.member.repository;
 
-import com.example.bookrecommandations.user.domain.Survey;
+import com.example.bookrecommandations.member.domain.Survey;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface SurveyRepository  extends JpaRepository<Survey, Long> {
