+++ src/main/java/com/example/bookrecommandations/member/vo/Age.java
@@ -1,4 +1,4 @@
-package com.example.bookrecommandations.user.vo;
+package com.example.bookrecommandations.member.vo;
 
 public enum Age {
     TEENS,
