+++ src/main/java/com/example/bookrecommandations/member/vo/Gender.java
@@ -1,0 +1,6 @@
+package com.example.bookrecommandations.member.vo;
+
+public enum Gender {
+    MALE,
+    FEMALE
+}
