+++ src/main/java/com/example/bookrecommandations/member/vo/MemberStatus.java
@@ -1,4 +1,4 @@
-package com.example.bookrecommandations.user.vo;
+package com.example.bookrecommandations.member.vo;
 
 public enum MemberStatus {
     ACTIVE,
