+++ src/main/java/com/example/bookrecommandations/member/vo/Role.java
@@ -1,0 +1,6 @@
+package com.example.bookrecommandations.member.vo;
+
+public enum Role {
+    USER,
+    ADMIN
+}
