--- src/main/java/com/example/bookrecommandations/user/vo/Gender.java
@@ -1,6 +1,0 @@
-package com.example.bookrecommandations.user.vo;
-
-public enum Gender {
-    MALE,
-    FEMALE
-}
