--- src/main/java/com/example/bookrecommandations/user/vo/Role.java
@@ -1,6 +1,0 @@
-package com.example.bookrecommandations.user.vo;
-
-public enum Role {
-    USER,
-    ADMIN
-}
